completed delete_post function, also added steps to prevent users from forcing and deleting or editing other users posts

+++ forum/views.py
@@ -1,24 +1,27 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
-from django.contrib.auth.models import User
 from django.contrib.auth.decorators import login_required
-from django.core.validators import ValidationError
 from django.contrib import messages
 from .models import Forum
 from .forms import CreatePostForm
 
 
-import datetime
-
 
 def share_builds(request):
     """ A view to return share_builds page """
+    # posts varible returns all of the posts in the Forum models
+    # user variable returns the request.user converted to a string
     posts = Forum.objects.all()
-    context = {"posts": posts}
+    user = str(request.user)
+    context = {
+        "posts": posts,
+        "user": user
+        }
     return render(request, 'forum/share_builds.html', context)
 
-
+@login_required()
 def create_post(request):
-    """ Adds posts to the share builds page"""
+    """ A view to add posts to the share builds page"""
+    # form variable returns the CreatePostForm created in forms.py
     form = CreatePostForm()
     context = {
         "form": form
@@ -35,24 +38,32 @@ def create_post(request):
         else:
             messages.error(request, 'Failed to update post. Please esnure the form is valid.')
 
-
     return render(request, 'forum/create_post.html', context)
 
 
+@login_required()
 def edit_post(request, post_id):
-    """Edit the post and return to the share builds page"""
+    """A view to Edit the post and return to the share builds page"""
+    # post varible returns the id of the post you want to update
+    # user variable returns the request.user converted to a string
     post = get_object_or_404(Forum, pk=post_id)
-    if request.method == 'POST':
-        form = CreatePostForm(request.POST, request.FILES, instance=post)
-        if form.is_valid():
-            form.save()
-            messages.success(request, 'Successfully updated post!')
-            return redirect(reverse('share_builds'))
+    user = str(request.user)
+
+    if post.poster.username == user:
+        if request.method == 'POST':
+            form = CreatePostForm(request.POST, request.FILES, instance=post)
+            if form.is_valid():
+                form.save()
+                messages.success(request, 'Successfully updated post!')
+                return redirect(reverse('share_builds'))
+            else:
+                messages.error(request, 'Failed to update post. Please esnure the form is valid.')
         else:
-            messages.error(request, 'Failed to update post. Please esnure the form is valid.')
+            form = CreatePostForm(instance=post)
+            messages.info(request, f'You are editing {post.title}')
     else:
-        form = CreatePostForm(instance=post)
-        messages.info(request, f'You are editing {post.title}')
+        messages.warning(request, 'Not allowed to update other users posts! ')
+        return redirect('home')
 
     context = {
         'post': post,
@@ -62,16 +73,24 @@ def edit_post(request, post_id):
     return render(request, 'forum/edit_post.html', context)
 
 
+@login_required()
 def delete_post(request, post_id):
+    """A view to Delete the post and return to the share builds page"""
+    # post varible returns the id of the post you want to update
+    # user variable returns the request.user converted to a string
     post = get_object_or_404(Forum, pk=post_id)
+    user = str(request.user)
 
     context = {
         'post': post,
     }
-
-    if request.method == 'POST':
-        post.delete()
-        messages.success(request,'Successfully deleted post!')
-        return redirect(reverse('share_builds'))
+    if post.poster.username == user:
+        if request.method == 'POST':
+            post.delete()
+            messages.success(request, 'Successfully deleted post!')
+            return redirect(reverse('share_builds'))
+    else:
+        messages.warning(request, 'Not allowed to delete other users posts! ')
+        return redirect('home')
     
     return render(request, 'forum/delete_post.html', context)
